==> 02_general_rules.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import json
import numpy as np
from matplotlib import pyplot as plt


pd.options.display.float_format = "{:,.2f}".format
sns.set_theme(style='whitegrid')

# Config Session

PATH_RULES = 'data/rules_v8.json'

PROTEINS = {
    1: 'CARNES',
    2: 'CARNES ULTRAPROCESSADAS',
    3: 'PESCADOS',
    4: 'OVOS',
    5: 'SUBSTITUTOS DE CARNE'
}

CODES = {
    11: 'CARNE BOVINA',
    12: 'CARNE SUINA',
    13: 'OUTRAS CARNES',
    14: 'AVES',
    15: 'CARNES SECAS E SALGADAS',
    16: 'VÍSCERAS',
    #17: 'PREPARAÇÕES QUE INCLUEM CARNES BOVINAS',
    #18: 'PREPARAÇÕES QUE INCLUEM OUTRAS CARNES',
    21: 'ALMÔNDEGAS, HAMBURGERS E NUGGETS',
    22: 'EMBUTIDOS, FRIOS E CONSERVAS',
    #23: 'PREPARAÇÕES QUE INCLUEM CARNES UP',	
    31: 'PESCADOS FRESCOS',
    32: 'PESCADOS EM CONSERVA',
    33: 'PESCADOS SALGADOS',
    #34: 'PREPARAÇÕES QUE INCLUEM PESCADOS',
    41: 'OVOS INTEIROS',
    #42: 'REFEIÇÕES QUE INCLUEM OVOS',
    51: 'PTS - CARNE DE SOJA',
    #52: 'PREPARAÇÕES QUE INCLUEM CARNE DE SOJA'
}

CORTES = {
    111: 'CORTE TRASEIRO',
    112: 'CORTE DIANTEIRO'
}

VIS_FIELDS = [
    'ano',
	'item',
	'un',
	'qt',
	'valor_un',
	'valor',
	'valor_real',
	'produtos',
	'un_adj',
    #'is_outlier_valor_un',
    #'is_outlier_qt',
    #'quartile_range_valor_un',
    #'quartile_range_qt'
]


CODE_GROUPS_WITH_MEDIAN_PRICE = [11, 12, 14, 16, 22, 51, 15, 33, 31, 32, 21, 13]

# List of products by group with unit_valor = 0
PRODUCTS_WITH_AVERAGE_PRICE = [
    # CARNE BOVINA
    "CARNE BOVINA DE 2ª, RESFRIADA/CONGELADA",
    "CARNE MOIDA",
    "CARNE BOVINA, PRE-PREAPADA, DIVERSOS CORTES",
    "CARNE BOVINA DE 1ª, RESFRIADA/CONGELADA",
    "CARNE BOVINA DE 3ª, RESFRIADA/CONGELADA",
    "CARNE BOVINA, ACEM, MOIDO",
	"CARNE BOVINA MOIDA",
    "CARNE BOVINA, COSTELA",
    "CARNE BOVINA, ACEM, SEM GORDURA",
    "CARNE BOVINA, MUSCULO, SEM GORDURA",
    "CARNE BOVINA, PALETA, SEM GORDURA",
    "CARNE BOVINA MOIDA MASTERBOI",
    "CARNE MOIDA COZIDA CONGELADA",
    "FRALDINHA",
    

    # AVES
    "FRANGO, PEITO, DIVERSOS CORTES, RESFRIADO/CONGELADO",
    "FRANGO PEITO, COM PELE",
    "FRANGO, COXA E SOBRECOXA, COM PELE",
    "FRANGO, COXA E SOBRECOXA, RESFRIADO/CONGELADO",
    "FRANGO INTEIRO, RESFRIADO/CONGELADO",
    "FRANGO INTEIRO, COM PELE",
    "FRANGO PEITO, SEM PELE",
    "FRANGO, FILE",
    "FRANGO CAIPIRA/SEMI CAIPIRA, RESFRIADO/CONGELADO",
    "COXA E SOBRECOXA DE FRANGO TEMPERADA E ASSADA SEM OSSO E SEM PELE EM CUBOS",
    "FRANGO, COXA E SOBRECOXA, SEM PELE",
    "PEITO DE FRANGO COZIDO TEMPERADO E DESFIADO",
    "FILEZINHO DE FRANGO EM TIRAS, TEMPERADO E GRELHADO",
    "FRANGO, COXA, RESFRIADO/CONGELADO",
    "FRANGO, PRE-PREAPADO, DIVERSOS CORTES",
    "FRANGO, COXA, COM PELE",
    "FRANGO, SOBRECOXA, RESFRIADO/CONGELADO",
    "FRANGO INTEIRO, SEM PELE",
    "FILE DE FRANGO DESFIADO COZIDO CONGELADO",
    "FRANGO, ASA, RESFRIADO/CONGELADO",
    "FRANGO, COXA DA ASA, RESFRIADO/CONGELADO",
    "FRANGO, COXA DA ASA",
    "FRANGO, SOBRECOXA, COM PELE",
    "FRANGO CAIPIRA",
    "FILE DE FRANGO GRELHADO CONGELADO",
    "FRANGO, ASA, COM PELE",
    "FRANGO, COXA, SEM PELE",
    "FRANGO, SOBRECOXA, SEM PELE",
    "FILE DE SASSAMI",
    "CARNE DE PERU, RESFRIADO/CONGELADO",
    "CARNE DE PATO, RESFRIADO/CONGELADO",
    "FRANGO SEMI-CAIPIRA",
    "CARNE DE AVESTRUZ",
    "PERU, CONGELADO",
    "AVESTRUZ",
    "PATO",
    
	# EMBUTIDOS, FRIOS E CONSERVAS
    "SALSICHA",
    "LINGUICA DE PORCO",
    "LINGUICA CALABRESA",
    "APRESUNTADO",
    "SALSICHA, DIVERSOS SABORES",
    "LINGUICA, DIVERSOS SABORES",
    "MORTADELA",
    "MORTADELA, DIVERSOS SABORES",
    "BACON",
    "PRESUNTO",
    "CARNE BOVINA, EM CONSERVA",
    "SALSICHA EM CONSERVA",
    "PRESUNTO, DIVERSOS SABORES",
    "SALSICHA DE FRANGO",
    "SALSICHA DE PERU",
    "SALSICHA, DIVERSOS SABORES, COM OU SEM MOLHO",
    "LINGUICA DE FRANGO",
    "FRANGO DESFIADO, COM OU SEM MOLHO, EM CONSERVA",
    "MORTADELA DE FRANGO",
    "SALAME",
    "CARNE BOVINA AO MOLHO COM LEGUMES, ENLATADA",
    "CARNE EM CONSERVA",
    "PASTEL ASSADO DE PRESUNTO E QUEIJO",
    "FRANGO DESFIADO AO MOLHO DE LEGUMES, ENLATADO",
    "SALSICHA AO MOLHO COM ERVILHA E MILHO (EM LATA)",
    
	# CARNE SUÍNA
    'TOUCINHO',
    'PORCO, BISTECA',
    'PORCO, LOMBO',
    'PORCO, MOIDO',
    'PORCO, PERNIL',
    'PORCO, COSTELA',
    'CARNE SUINA, DIVERSOS CORTES',
    
	#
    'CARNE BOVINA, CHARQUE',
    'CARNE BOVINA, SECA',
    'CARNE DE SOL',
    'CARNE SECA/CHARQUE/JERKED BEEF',
    'PORCO, RABO, SALGADO',
    'PORCO, ORELHA, SALGADA',
    
    
	# PTS
    'PROTEINA TEXTURIZADA DE SOJA',
    
	# 33
	'CAMARAO SECO',
    'PEIXE, AGUA DOCE, SECO',
    'PEIXE, AGUA SALGADA, SECO',
    'FRUTOS DO MAR, SECO',
    'BACALHAU SALGADO',
    'SARDINHA SECA',
    
	# 31
    "PESCADA, FILE",
    "PEIXE, AGUA DOCE, CONGELADO",
    "PEIXE, AGUA SALGADA, CONGELADO",
    "PEIXE, AGUA SALGADA, FRESCO",
    "PEIXE, AGUA DOCE, FRESCO",
    "PEIXE TAMBACU",
    "MERLUZA, FILE",
    "TILAPIA, FILE",
    "PINTADO",
    "CORVINA DO MAR",
    "FILE DE PEIXE",
    "CAMARAO FRESCO",
    "PEIXE PANGACO",
    "PEIXE SEM ESPINHA",
    "PEIXE MOQUEADO",
    "CORIMBA",
    "CORVINA DE AGUA DOCE",
    "SARDINHA, INTEIRA, CRUA",
    "ATUM FRESCO",
    "FRUTOS DO MAR, CONGELADO",
    "FRUTOS DO MAR, FRESCO",
    "PEIXE COZIDO",
    "PEIXE ASSADO",
    "CACAO, POSTA",
    "PESCADA BRANCA",
    "PESCADINHA",
    "ANJO, FILE",
    "DOURADA DE AGUA DOCE, FRESCA",
    "MAPARA, FILE",
    "ABADEJO, FILE, CONGELADO",
    "POLACA, FILE",
    "TILAPIA, INTEIRA",
    "TUCUNARE, FILE, CONGELADO, CRU",
    "SIRI",
    "OSTRA",
    
	# 32
	'ATUM, CONSERVA EM OLEO',
    'SARDINHA EM OLEO',
    'SARDINHA EM CONSERVA',
	'SARDINHA, CONSERVA EM AGUA',
	'SARDINHA, CONSERVA EM OLEO',
	'SARDINHA, CONSERVA EM MOLHO DE TOMATE',
	'ANCHOITA, CONSERVA EM MOLHO DE TOMATE',
	'PEIXE, EM CONSERVA',


	# 16
	'CARNE BOVINA, FIGADO',
	'CARNE BOVINA, BUCHO',
	'FRANGO, FIGADO',
	'FRANGO, CORACAO',
	'FRANGO, MOELA',
	'CARNE BOVINA, LINGUA',
	'CARNE BOVINA, VISCERAS',
	'FRANGO, VISCERAS, RESFRIADO/CONGELADO',
    

	# 21
    'ALMONDEGA AO MOLHO, ENLATADA',
    'CARNE BOVINA, ALMONDEGAS, CRUAS',
    'NUGGETS, DIVERSOS SABORES',
    'ALMONDEGA, DIVERSOS SABORES, RESFRIADA/CONGELADA',
    'ALMONDEGA DE PEIXE',
    'ALMONDEGA AO MOLHO, EM CONSERVA',
    'HAMBURGUER, BOVINO',
    'HAMBURGUER MISTO',
    'HAMBURGUER DE FRANGO',
    
	# 13
    'CARNE DE CAPRINO/CARNE DE BODE/CARNE DE OVELHA, RESFRIADA/CONGELADA',
	'CARNE DE CAPRINO/CARNE DE BODE',
	'CARNE DE OVINO',
	'CARNE DE CACA, RESFRIADA/CONGELADA',
	'CARNE DE CACA'
]

version = '9'

# Input Data

df_units_products = pd.read_csv('data/stat_unidades_por_produto.csv')
df_dataset_beef = pd.read_parquet('data/banco_carnes_un_v2.parquet')
df_valores_carne = pd.read_excel('data/valores_medios_ano.xlsx')
df_valores_carne_f = pd.read_csv('data/valores_medios_ano_filled.csv')


with open(PATH_RULES, 'r') as f: rules = json.load(f)

# Add Properties

df_dataset_beef['unit_kg'] = 0
df_dataset_beef['price_kg'] = 0
df_dataset_beef['median_price'] = 0
df_dataset_beef['rule_apply'] = 0
df_dataset_beef['outlier'] = 0
df_dataset_beef['rule_version'] = ''
df_dataset_beef['rule_id'] = ''

# Helpers


def set_outlier_property(df, col):
    def classify_quartile(value, q1, q3):
        q1_q2 = (q1 + q3) / 2
        if value <= q1:
            return f'Q1 (<= {q1:.1f})'
        elif q1 < value <= q1_q2:
            return f'Q1-Q2 ({q1:.1f} - {q1_q2:.1f})'
        elif q1_q2 < value <= q3:
            return f'Q3-Q4 ({q1_q2:.1f} - {q3:.1f})'
        else:
            return f'Q4 (> {q3:.1f})'

    q1 = df[col].quantile(0.10)
    q3 = df[col].quantile(0.95)
    iqr = q3 - q1

    lower_bound = q1 - 1.5 * iqr
    upper_bound = q3 + 1.5 * iqr

    df[f'is_outlier_{col}'] = (~df[col].between(lower_bound, upper_bound)).astype(int)
    df[f'quartile_range_{col}'] = df[col].apply(lambda x: classify_quartile(x, q1, q3))
    return df

# Fix initial values


# ajusta valores nulos do campo valor_un
df_dataset_beef.loc[
    (df_dataset_beef['valor_un'].isna()) & 
    (df_dataset_beef['qt'] >= 1) &
    (df_dataset_beef['codigo'] == 41) &
    (df_dataset_beef['un_adj'] == 'DZ'),
    'valor_un'
] = df_dataset_beef['valor'].div(df_dataset_beef['qt'])

df_dataset_beef = df_dataset_beef.fillna(0)

# Identifies quartis and intervals
this process add a property containing classes of quartis

df_dataset_beef_cls = df_dataset_beef.groupby(['produtos', 'un_adj'], group_keys=False)\
    .apply(lambda group: set_outlier_property(group, 'valor_un'))

df_dataset_beef_cls = df_dataset_beef_cls.groupby(['produtos', 'un_adj'], group_keys=False)\
    .apply(lambda group: set_outlier_property(group, 'qt'))

# Checking records

# Filtrar os dados com um único filtro consistente
filtered_data = df_dataset_beef_cls.query('codigo == 13 and un_adj == "KG"')
total = len(filtered_data)  # Total de registros no filtro

# Groupby com cálculos ajustados
result = (
    filtered_data
    .groupby(['un_adj', 'quartile_range_valor_un', 'is_outlier_valor_un'])
    .agg(
        count=('valor_un', 'count'),
        q1=('valor_un', lambda x: x.quantile(0.25)),  # 1º quartil
        q2=('valor_un', 'median'),  # Mediana (2º quartil)
        q3=('valor_un', lambda x: x.quantile(0.75)),  # 3º quartil
        min=('valor_un', 'min'),  # Valor mínimo
        max=('valor_un', 'max'),  # Valor máximo
    )
    .reset_index()
)

result['count_percent'] = result['count'].div(result['count'].sum())

result

# Add Reference Values from Auxiliary
This code section adds a column containing the reference values verified by the team.

df_dataset_beef_cls['valor_un_ref'] = 0

ref_names = [
    'CARNE BOVINA',
    'ACÉM', 
    'CHARQUE PA',
    'COSTELA',
    'PONTA DE AGULHA',
	'OVO'
    #'QUARTO DIANTEIRO COM OSSO',
    #'QUARTO TRASEIRO COM OSSO'
]

# "corte" significa corte traseiro (=111) ou dianteiro (=112)

for ref_name in ref_names:
	for year in list(range(2013, 2023)):

		ref_val = df_valores_carne_f.loc[
			df_valores_carne_f['item'].str.contains(ref_name)
		][f'valor_{str(year)}'].values[0]

		if ref_name == 'ACÉM':
			df_dataset_beef_cls.loc[
				(df_dataset_beef_cls['ano'] == year) &
				(df_dataset_beef_cls['codigo'] == 11) &
				(df_dataset_beef_cls['item'].str.contains('ACEM', na=False)),
				'valor_un_ref'
			] = ref_val

		else:
			df_dataset_beef_cls.loc[
				(df_dataset_beef_cls['ano'] == year) &
				(df_dataset_beef_cls['item'].str.contains(ref_name, na=False)),
				'valor_un_ref'
			] = ref_val



# Add median value by product

# general median value for each group by year
for year in list(range(2013, 2023)):
	for code, product in CODES.items():

		median_price = df_dataset_beef_cls.query('ano == @year and codigo == @code and un_adj == "KG"')['valor_un'].median()

		df_dataset_beef_cls.loc[
			(df_dataset_beef_cls['ano'] == year) &
			(df_dataset_beef_cls['codigo'] == code),
			'median_price'
		] = median_price


# updates the group average price with the detailed average price when available
median_price_products = df_dataset_beef_cls.loc[
    (df_dataset_beef_cls['codigo'].isin(CODE_GROUPS_WITH_MEDIAN_PRICE)) &
    (df_dataset_beef_cls['un_adj'] == 'KG') &
    (df_dataset_beef_cls['valor_un'] != 0) &
    (df_dataset_beef_cls['is_outlier_qt'] == 0) &
    (df_dataset_beef_cls['valor'] != 0) &
    (df_dataset_beef_cls['item'].isin(PRODUCTS_WITH_AVERAGE_PRICE))
].groupby(by='item')['valor_un'].median().reset_index().rename(columns={'valor_un': 'valor_un_median'})

# set detailed median price
for index, row in median_price_products.iterrows():
	df_dataset_beef_cls.loc[
		(df_dataset_beef_cls['codigo'].isin(CODE_GROUPS_WITH_MEDIAN_PRICE)) &
		(df_dataset_beef_cls['item'] == row['item']),
		'median_price'
	] = row['valor_un_median']



# Apply rule on 1% higher qt

df_dataset_beef_cls['qt_origin'] = df_dataset_beef_cls['qt']

for code, product in CODES.items():
	
	df_product = df_dataset_beef_cls.query('codigo == @code').sort_values(by='qt', ascending=False)

	count_vals = df_product.shape[0]
	count_lim = int(count_vals/100)


	# adj 1% of higher qt
	df_product_adj = df_product.iloc[0: count_lim].query('median_price > 0')

	df_dataset_beef_cls.loc[df_product_adj.index, 'valor_un'] = df_product_adj['median_price']
	df_dataset_beef_cls.loc[df_product_adj.index, 'qt'] = df_product_adj['valor'].div(df_product_adj['median_price'])


	

# Applying Rules



# iterate over rules by products
for k, v in rules.items():
    # get rule dict
    for rule in v:
        filter_condition = rule['filter']
        unit_kg_expression = rule['unit_kg']
        price_kg_expression = rule['price_kg']

        rule_version = rule['version']
        rule_id = rule['id']

        # filter rows based on the filter_condition
        filtered_rows = df_dataset_beef_cls.query(filter_condition)

        # Apply unit_kg expression, rule_id, price_kg expression, and rule_version
        df_dataset_beef_cls.loc[filtered_rows.index, 'unit_kg'] = filtered_rows.eval(unit_kg_expression)
        df_dataset_beef_cls.loc[filtered_rows.index, 'rule_id'] = rule_id
        df_dataset_beef_cls.loc[filtered_rows.index, 'rule_version'] = rule_version
        df_dataset_beef_cls.loc[filtered_rows.index, 'rule_apply'] = 1
        df_dataset_beef_cls.loc[filtered_rows.index, 'price_kg'] = filtered_rows.eval(price_kg_expression)
        

df_dataset_beef_cls.replace([np.inf, -np.inf], 0, inplace=True)
df_dataset_beef_cls = df_dataset_beef_cls.fillna(0)


## Check Results

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 41) &
    (df_dataset_beef_cls['rule_apply'] == 1)
][VIS_FIELDS + ['price_kg', 'rule_apply']].sort_values('price_kg', ascending=True)

# Set Outliers

set outliers and export

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 41) &
    ((df_dataset_beef_cls['price_kg'] > 25) | (df_dataset_beef_cls['price_kg'] < 3)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 11) &
    ((df_dataset_beef_cls['price_kg'] > 60) | (df_dataset_beef_cls['price_kg'] < 3)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 14) &
    ((df_dataset_beef_cls['price_kg'] > 38) | (df_dataset_beef_cls['price_kg'] < 3)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 22) &
    ((df_dataset_beef_cls['price_kg'] > 40) | (df_dataset_beef_cls['price_kg'] < 2.55)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 12) &
    ((df_dataset_beef_cls['price_kg'] > 40) | (df_dataset_beef_cls['price_kg'] < 2.55)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 51) &
    ((df_dataset_beef_cls['price_kg'] > 43) | (df_dataset_beef_cls['price_kg'] < 3)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 15) &
    ((df_dataset_beef_cls['price_kg'] > 60) | (df_dataset_beef_cls['price_kg'] < 4)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 31) &
    ((df_dataset_beef_cls['price_kg'] > 45) | (df_dataset_beef_cls['price_kg'] < 3)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 33) &
    ((df_dataset_beef_cls['price_kg'] > 45) | (df_dataset_beef_cls['price_kg'] < 3)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 32) &
    ((df_dataset_beef_cls['price_kg'] > 45) | (df_dataset_beef_cls['price_kg'] < 5)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 16) &
    ((df_dataset_beef_cls['price_kg'] > 45) | (df_dataset_beef_cls['price_kg'] < 3)),
    'outlier'
] = 1

df_dataset_beef_cls.loc[
	(df_dataset_beef_cls['codigo'] == 21) &
    ((df_dataset_beef_cls['price_kg'] > 60) | (df_dataset_beef_cls['price_kg'] < 3)),
    'outlier'
] = 1

# Apply special rule for outliers

sp_rule_1 = df_dataset_beef_cls.query('rule_apply == 1 and un_adj == "KG" and price_kg == 0 and qt == 0 and valor == valor_un and valor.between(3, 55)')
sp_rule_2 = df_dataset_beef_cls.query('rule_apply == 1 and un_adj == "KG" and price_kg == 0 and valor_un == 0 and valor != 0 and qt != 0 and (valor/qt) > 2 and (valor/qt) < 55')

df_dataset_beef_cls.loc[sp_rule_1.index, 'price_kg'] = sp_rule_1['valor']
df_dataset_beef_cls.loc[sp_rule_1.index, 'unit_kg'] = 1
df_dataset_beef_cls.loc[sp_rule_1.index, 'rule_id'] = 'ALL_SP_1'         
df_dataset_beef_cls.loc[sp_rule_1.index, 'rule_apply'] = 1                                                                    
																			 
df_dataset_beef_cls.loc[sp_rule_2.index, 'price_kg'] = sp_rule_2['valor'].div(sp_rule_2['qt'])
df_dataset_beef_cls.loc[sp_rule_2.index, 'unit_kg'] = sp_rule_2['qt']
df_dataset_beef_cls.loc[sp_rule_2.index, 'rule_id'] = 'ALL_SP_2'  
df_dataset_beef_cls.loc[sp_rule_2.index, 'rule_apply'] = 1


# Export

df_dataset_beef_cls['produtos'] = df_dataset_beef_cls['produtos'].astype(str)
df_dataset_beef_cls['dados.CNPJ.CPF'] = df_dataset_beef_cls['dados.CNPJ.CPF'].astype(str)
df_dataset_beef_cls['dados.emissao'] = df_dataset_beef_cls['dados.emissao'].astype(str)
df_dataset_beef_cls['un'] = df_dataset_beef_cls['un'].astype(str)

# drop fields
df_dataset_beef_cls.drop(['is_outlier_valor_un','quartile_range_valor_un','is_outlier_qt','quartile_range_qt'],axis=1,inplace=True)

df_dataset_beef_cls.to_parquet(f'data/banco_carnes_un_v{version}_ed.parquet')

# Compute Stat

df_stat_rule = df_dataset_beef_cls.groupby(['produtos','rule_apply']).agg(
    count=('rule_apply', 'count')
).reset_index()



df_stat_rule['percent'] = df_stat_rule['count'] / df_stat_rule.groupby('produtos')['count'].transform('sum') * 100
df_stat_rule.to_csv('data/stat_rule_apply.csv')

df_stat_rule

df_dataset_beef_cls.loc[
    (df_dataset_beef_cls['codigo'] == 51) & (df_dataset_beef_cls['rule_apply'] == 1)
].groupby('outlier').count()

df_dataset_beef_cls.groupby(by=['rule_apply']).agg(	
    count=('rule_apply', 'count'),
    percent_prod=('rule_apply', lambda x: len(x) / len(df_dataset_beef_cls) * 100)
).reset_index()
