==> 05_rename_tables.ipynb <==
# Rename Tables and columns

import pandas as pd

from glob import glob

# Config

PATH_TABLES = 'data/comparacoes'

FIELDS = [
    'ano', 'item',
    'qt_nova', 'unit_kg',
    'valor_kg', 'price_kg', 
    'valor_novo', 'valor_novo_s'
]

NEW_FIELDS = [
    'ano', 'item', 
    'qt_kg_amostra', 'qt_kg_estimado',
    'valor_un_kg_amostra', 'valor_un_kg_estimado',
    'valor_total_kg_amostra', 'valor_total_kg_estimado'
]

# Rename

d = []
for path in glob(f'{PATH_TABLES}/produto_14*'):
	code = int(path.split('_')[1])
	
	df_comp = pd.read_csv(path)[FIELDS].rename(columns=dict(zip(FIELDS, NEW_FIELDS)))
	df_comp['codigo'] = code

	d.append(df_comp)
	
result = pd.concat(d)


result.to_csv('data/14_table_comparasions.csv')

result
